+++ client/src/components/signup/willingOrganisation/stepOne.js
@@ -83,7 +83,7 @@ export default () => {
                   </div>
               </div>
               
-              <button style={{ background: "#ffc107", borderWidth: "2px", color: "6c7a86", borderRadius: "1.2em", padding: "4px",height:"40px",width:"100px"}}   type="submit" >Save</button>
+              <button style={{ background: "#33c3f0", borderWidth: "2px", color: "6c7a86", borderRadius: "1.2em", padding: "4px",height:"40px",width:"100px"}}   type="submit" >Save</button>
               
            </form>   
     </div>
