+++ server/app.js
@@ -12,7 +12,7 @@ const activitiesRouter = require("./routes/activities");
 const feedRouter = require("./routes/feed");
 const requestRouter = require("./routes/request");
 const signupRouter = require("./routes/sign_up");
-
+const reqRouter = require("./routes/donation-need-routes/request")
 
 const route1 = require('./routes/userSignup');
 const route2 = require('./routes/orgSignup');
@@ -20,7 +20,7 @@ const editUser = require('./routes/api.js');
 
 
 const { SESSION_SECRET } = require("./secretConfig");
-const sgMail = require('@sendgrid/mail')
+
 require("dotenv").config();
 const User = require("./models/person");
 var string = require("string-sanitizer");
@@ -37,7 +37,7 @@ mongoose
     useNewUrlParser: true,
     useCreateIndex: true,
     useUnifiedTopology: true,
-     dbName: DB_NAME,
+     dbName: "finaldb",
     w: "majority",
   })
   .catch((err) => {
@@ -88,192 +88,24 @@ app.use(
   })
 );
 
-//app.post('/sign_up', (req, res) => {
-  //  console.log('Hello')
-//})
 
 
 
-
-
-
-app.post("/sign_up/", async (req, res) => { // finally url will be "/sign_up/" (as the previous one used)
-    console.log('Okay we can reach here')
-
-
-    bcrypt.hash(req.body.password, 10, (err, hash) => {
-        if (err) {
-            console.log("Password can not be encrypted");
-        }
-        user = new User({
-            name: req.body.name,
-            contactNumber: req.body.contactNumber,
-            email: req.body.email,
-            password: hash
-        });
-        user.save().then(() => {
-            console.log(user);
-            
-            res.sendStatus(200);
-        })
-        .catch(err => res.status(500).send(err));
-
-    });
-
-    // res.redirect('/willingorganisationsignupstep2')
-});
-
-var OTP;
-app.post('/verify-email', (req, res) => {
-    console.log('Hey there')
-    sgMail.setApiKey(process.env.API_KEY)
-    stringx = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
-    OTP = '';
-    var len = stringx.length;
-    for (let i = 0; i < 6; i++) {
-        OTP += stringx[Math.floor(Math.random() * len)];
-    }
-
-
-    mail = 'Your OTP is ' + OTP;
-    const msg = {
-        to: user.email, // Change to your recipient
-        from: 'muckinverify@gmail.com', // Change to your verified sender
-        subject: 'OTP',
-        text: mail,
-
-    }
-    sgMail
-        .send(msg)
-        .then(() => {
-            console.log('Email sent')
-        })
-        .catch((error) => {
-            console.error(error + ' Cannot send')
-        })
-    console.log('OTP ' + OTP)
-    res.redirect('/willingorganisationsignupstep2')
-})
-
-app.post('/verifyemail', (req, res) => {
-    otpReq = string.sanitize(req.body.otp);
-    if (otpReq === OTP) {
-        user.isVerifiedEmail = true;
-    }
-    user.save().then(() => {
-        console.log(user)
-    })
-    res.redirect('/willingorganisationsignupstep2')
-    console.log(otpReq);
-
-    console.log("OTP " + OTP);
-
-
-})
 
 // Routes START
 app.use("/user", userRouter); // login, logout
-app.use("/sign_up", signupRouter); // sign_up individual and organisation
+app.use(signupRouter); // sign_up individual and organisation
 app.use("/activities", activitiesRouter); // image, update-details, delete-details
 app.use("/requests", requestRouter); // /new request
 app.use("/feeds", feedRouter); // /get feeds
 // Routes END
-app.use("/org", signupRouter);
+
 app.use('/editUser',editUser); // edit user profile
 app.use('/api1', route1);  // signup user looking for help
 app.use('/api2',route2); // signup org looking for help
 
-app.use('/activity',activRouter);
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-app.post("/org", (req, res) => { // finally url will be "/sign_up/org" (as the previous one used)
-    user.Name_of_organisation = string.sanitize(req.body.nameorganisation);
-    user.Address_of_organisation = string.sanitize(
-        req.body.addressorganisation
-    );
-    user.License_number = string.sanitize(req.body.license);
-    user.Type_of_organisation = string.sanitize(req.body.typeorganisation);
-    user.Description_of_organisation = string.sanitize(
-        req.body.describe
-    );
-    user.Volunteers_number = string.sanitize(req.body.numbervolun);
-    user.Type_of_help = string.sanitize(req.body.help);
-    user.Open_for_volunteers = string.sanitize(req.body.isopen);
-    // console.log(typeof(req.body.Address_of_organisation))
-    // console.log(typeof(req.body.Volunteers_number))
-
-    User.updateOne(
-        { _id: user._id },
-        {
-            name: user.name,
-            email: user.email,
-            contactNumber: user.contactNumber,
-            password: user.password,
-            Name_of_organisation: user.Name_of_organisation,
-            Address_of_organisation: user.Address_of_organisation,
-            License_number: user.License_number,
-            Type_of_organisation: user.Type_of_organisation,
-            Description_of_organisation: user.Description_of_organisation,
-            Volunteers_number: user.Volunteers_number,
-            Type_of_help: user.Type_of_help,
-            Open_for_volunteers: user.Open_for_volunteers,
-        }
-    ).then(() => {
-        console.log(user)
-    });
-
-
-
-    res.redirect("/willingorganisationsignupstep3")
-});
-
-app.post("/accept", (req, res) => {
-    res.redirect('/')
-})
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+app.use('/activity',activitiesRouter);
+app.use(reqRouter)
